fix(wezterm): split setting
