+++ apps/core/wezterm/wezterm/appearance.lua
@@ -1,88 +1,5 @@
 local wezterm = require("wezterm")
-
--- The filled in variant of the < symbol
-local SOLID_LEFT_ARROW = ""
-
--- The filled in variant of the > symbol
-local SOLID_RIGHT_ARROW = utf8.char(0xe0b0)
-
-local color_set = {
-    bg = "#282a36",
-    fg = "#f8f8f2",
-    bg_tab = "#44475a",
-    fg_tab = "#f8f8f2",
-    bg_active = "#6272a4",
-    fg_active = "#f8f8f2",
-}
-
-local function tab_title(tab_info)
-    local title = tab_info.tab_title
-    -- if the tab title is explicitly set, take that
-    if title and #title > 0 then
-        return title
-    end
-    -- Otherwise, use the title from the active pane
-    -- in that tab
-    return tab_info.active_pane.title
-end
-
-wezterm.on("format-tab-title", function(tab, tabs, panes, config, hover, max_width)
-    local edge_background = color_set.bg_tab
-    local right_edge_background = nil
-    local background = color_set.bg_tab
-    local foreground = color_set.fg_tab
-    local zoomed = ""
-    local extra_pad = 0
-    if tab.active_pane.is_zoomed then
-        zoomed = "[Z] "
-        extra_pad = 4
-    end
-    if tab.is_active then
-        if tab.tab_index == (#tabs - 1) then
-            edge_background = color_set.bg
-        end
-        background = color_set.bg_active
-        foreground = color_set.fg_active
-    elseif not tab.is_active then
-        if tab.tab_index == (#tabs - 1) then
-            edge_background = color_set.bg
-            SOLID_RIGHT_ARROW = utf8.char(0xe0b0)
-        elseif (tab.tab_index + 1) <= (#tabs - 1) and not tabs[tab.tab_index + 2].is_active then
-            edge_background = color_set.bg_tab
-            right_edge_background = "#7aa2f7"
-            SOLID_RIGHT_ARROW = utf8.char(0xe0b1)
-        else
-            edge_background = color_set.bg_active
-            SOLID_RIGHT_ARROW = utf8.char(0xe0b0)
-        end
-        background = color_set.bg_tab
-        foreground = color_set.fg_tab
-    elseif hover then
-        background = "#3b3052"
-        foreground = "#909090"
-    end
-
-    local edge_foreground = background
-    if right_edge_background == nil then
-        right_edge_background = background
-    end
-    -- ensure that the titles fit in the available space,
-    -- and that we have room for the edges.
-    local title = tab_title(tab)
-    title = " " .. zoomed .. tab.tab_index .. " " .. wezterm.truncate_right(title, max_width - 5 - extra_pad) .. " "
-
-    return {
-        { Background = { Color = edge_background } },
-        { Foreground = { Color = edge_foreground } },
-        { Text = SOLID_LEFT_ARROW },
-        { Background = { Color = background } },
-        { Foreground = { Color = foreground } },
-        { Text = title },
-        { Background = { Color = edge_background } },
-        { Foreground = { Color = right_edge_background } },
-        { Text = SOLID_RIGHT_ARROW },
-    }
-end)
+local color_set = require("color-set")
 
 local config = {
     window_close_confirmation = "AlwaysPrompt",
