+++ apps/core/wezterm/wezterm/color-set.lua
@@ -1,0 +1,8 @@
+return {
+    bg = "#282a36",
+    fg = "#f8f8f2",
+    bg_tab = "#44475a",
+    fg_tab = "#f8f8f2",
+    bg_active = "#6272a4",
+    fg_active = "#f8f8f2",
+}
